# contracts/openapi.bundle.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "YuLab foundation API",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000"
    }
  ],
  "paths": {
...
    "/api/v1/capabilities": {
      "get": {
        "operationId": "getCapabilities",
        "responses": {
          "200": {
            "description": "Capabilities",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Capabilities"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/captures": {
      "post": {
        "operationId": "createCapture",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaptureRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Capture accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaptureJob"
                }
              }
            }
          },
          "422": {
            "description": "Invalid capture request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "operationId": "getCaptureJob",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Capture job",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaptureJob"
                }
              }
            }
          },
          "404": {
            "description": "Unknown capture job",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/cancel": {
      "post": {
        "operationId": "cancelCaptureJob",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cancellation requested",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaptureJob"
                }
              }
            }
          },
          "404": {
            "description": "Unknown capture job",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/artifacts/{artifact_id}": {
      "get": {
        "operationId": "getCaptureArtifact",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          },
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Published capture artifact",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Unknown or unpublished artifact",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Id": {
        "type": "string",
        "pattern": "^[a-zA-Z0-9][a-zA-Z0-9_-]{0,95}$",
        "maxLength": 96
      },
      "Hash": {
        "type": "string",
        "pattern": "^[a-f0-9]{64}$",
        "minLength": 64,
        "maxLength": 64
      },
      "Version": {
        "type": "string",
        "minLength": 1,
        "maxLength": 64
      },
      "Vec3": {
        "type": "array",
        "items": {
          "type": "number",
          "minimum": -1000000000,
          "maximum": 1000000000
        },
        "minItems": 3,
        "maxItems": 3
      },
...
      "Units": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "length": {
            "const": "m"
          },
          "temperature": {
            "const": "K"
          },
          "angle": {
            "const": "rad"
          },
          "time": {
            "const": "s"
          },
          "world_frame": {
            "const": "east-up-south"
          },
          "matrix_layout": {
            "const": "row-major"
          }
        },
        "required": [
          "length",
          "temperature",
          "angle",
          "time",
          "world_frame",
          "matrix_layout"
        ]
      },
...
      "Feature": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Id"
          },
          "type": {
            "enum": [
              "canyon",
              "alpine",
              "islands",
              "coast",
              "aerodrome",
              "harbor"
            ]
          },
          "center_m": {
            "$ref": "#/components/schemas/Vec3"
          },
          "extent_m": {
            "type": "array",
            "items": {
              "type": "number",
              "minimum": 0.01,
              "maximum": 100000
            },
            "minItems": 3,
            "maxItems": 3
          },
          "seed": {
            "type": "integer",
            "minimum": 0,
            "maximum": 4294967295
          }
        },
        "required": [
          "id",
          "type",
          "center_m",
          "extent_m",
          "seed"
        ]
      },
      "WorldSpec": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "schema_version": {
            "const": "lab.v1"
          },
          "kind": {
            "const": "WorldSpec"
          },
          "world_id": {
            "$ref": "#/components/schemas/Id"
          },
          "seed": {
            "type": "integer",
            "minimum": 0,
            "maximum": 4294967295
          },
          "units": {
            "$ref": "#/components/schemas/Units"
          },
          "extent_m": {
            "type": "array",
            "items": {
              "type": "number",
              "minimum": 1,
              "maximum": 100000
            },
            "minItems": 3,
            "maxItems": 3
          },
          "origin_m": {
            "$ref": "#/components/schemas/Vec3"
          },
          "chunk_size_m": {
            "type": "number",
            "minimum": 1,
            "maximum": 4096
          },
          "generator_version": {
            "$ref": "#/components/schemas/Version"
          },
          "field_version": {
            "$ref": "#/components/schemas/Version"
          },
          "features": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Feature"
            },
            "minItems": 0,
            "maxItems": 1000
          },
          "instances": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Instance"
            },
            "minItems": 0,
            "maxItems": 100000
          }
        },
        "required": [
          "schema_version",
          "kind",
          "world_id",
          "seed",
          "units",
          "extent_m",
          "origin_m",
          "chunk_size_m",
          "generator_version",
          "field_version",
          "features",
          "instances"
        ]
      },
...
      "RigSpec": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "schema_version": {
            "const": "lab.v1"
          },
          "kind": {
            "const": "RigSpec"
          },
          "rig_id": {
            "$ref": "#/components/schemas/Id"
          },
          "calibration_version": {
            "$ref": "#/components/schemas/Version"
          },
          "T_world_from_rig": {
            "$ref": "#/components/schemas/Transform"
          },
          "sensors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SensorRig"
            },
            "minItems": 3,
            "maxItems": 3,
            "allOf": [
              {
                "contains": {
                  "properties": {
                    "modality": {
                      "const": "rgb"
                    }
                  }
                },
                "minContains": 1,
                "maxContains": 1
              },
              {
                "contains": {
                  "properties": {
                    "modality": {
                      "const": "ir"
                    }
                  }
                },
                "minContains": 1,
                "maxContains": 1
              },
              {
                "contains": {
                  "properties": {
                    "modality": {
                      "const": "lidar"
                    }
                  }
                },
                "minContains": 1,
                "maxContains": 1
              }
            ]
          },
          "timing_profile": {
            "enum": [
              "synchronized-static",
              "timed-scan"
            ]
          }
        },
        "required": [
          "schema_version",
          "kind",
          "rig_id",
          "calibration_version",
          "T_world_from_rig",
          "sensors",
          "timing_profile"
        ]
      },
      "EnvironmentSpec": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "schema_version": {
            "const": "lab.v1"
          },
          "kind": {
            "const": "EnvironmentSpec"
          },
          "environment_id": {
            "$ref": "#/components/schemas/Id"
          },
          "simulation_time_s": {
            "type": "number",
            "minimum": 0,
            "maximum": 1000000000
          },
          "solar_time_hour": {
            "type": "number",
            "minimum": 0,
            "maximum": 24
          },
          "latitude_rad": {
            "type": "number",
            "minimum": -1.5707963267948966,
            "maximum": 1.5707963267948966
          },
          "sun_direction_world": {
            "$ref": "#/components/schemas/Vec3"
          },
          "ambient_temperature_k": {
            "type": "number",
            "minimum": 100,
            "maximum": 500
          },
          "fog_extinction_per_m": {
            "type": "number",
            "minimum": 0,
            "maximum": 10
          },
          "rain_mm_per_h": {
            "type": "number",
            "minimum": 0,
            "maximum": 1000
          },
          "snow_mm_per_h": {
            "type": "number",
            "minimum": 0,
            "maximum": 1000
          },
          "wind_m_per_s": {
            "$ref": "#/components/schemas/Vec3"
          },
          "wetness": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          },
          "parameter_set_version": {
            "$ref": "#/components/schemas/Version"
          },
          "thermal_history": {
            "enum": [
              "equilibrated",
              "continued"
            ]
          },
          "thermal_state": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Artifact"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "schema_version",
          "kind",
          "environment_id",
          "simulation_time_s",
          "solar_time_hour",
          "latitude_rad",
          "sun_direction_world",
          "ambient_temperature_k",
          "fog_extinction_per_m",
          "rain_mm_per_h",
          "snow_mm_per_h",
          "wind_m_per_s",
          "wetness",
          "parameter_set_version",
          "thermal_history",
          "thermal_state"
        ]
      },
      "CapturePlan": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "schema_version": {
            "const": "lab.v1"
          },
          "kind": {
            "const": "CapturePlan"
          },
          "capture_id": {
            "$ref": "#/components/schemas/Id"
          },
          "sequence_id": {
            "$ref": "#/components/schemas/Id"
          },
          "world_sha256": {
            "$ref": "#/components/schemas/Hash"
          },
          "rig_sha256": {
            "$ref": "#/components/schemas/Hash"
          },
          "environment_sha256": {
            "$ref": "#/components/schemas/Hash"
          },
          "simulation_time_s": {
            "type": "number",
            "minimum": 0,
            "maximum": 1000000000
          },
          "geometry_policy": {
            "const": "fixed-sensor-geometry"
          },
          "quality_version": {
            "$ref": "#/components/schemas/Version"
          },
          "seed_channels": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "world": {
                "type": "integer",
                "minimum": 0,
                "maximum": 4294967295
              },
              "weather": {
                "type": "integer",
                "minimum": 0,
                "maximum": 4294967295
              },
              "rgb": {
                "type": "integer",
                "minimum": 0,
                "maximum": 4294967295
              },
              "ir": {
                "type": "integer",
                "minimum": 0,
                "maximum": 4294967295
              },
              "lidar": {
                "type": "integer",
                "minimum": 0,
                "maximum": 4294967295
              }
            },
            "required": [
              "world",
              "weather",
              "rgb",
              "ir",
              "lidar"
            ]
          },
          "modalities": {
            "type": "array",
            "items": {
              "enum": [
                "rgb",
                "ir",
                "lidar"
              ]
            },
            "minItems": 1,
            "maxItems": 3,
            "uniqueItems": true
          }
        },
        "required": [
          "schema_version",
          "kind",
          "capture_id",
          "sequence_id",
          "world_sha256",
          "rig_sha256",
          "environment_sha256",
          "simulation_time_s",
          "geometry_policy",
          "quality_version",
          "seed_channels",
          "modalities"
        ]
      },
      "CaptureRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "schema_version": {
            "const": "lab.v1"
          },
          "kind": {
            "const": "CaptureRequest"
          },
          "world": {
            "$ref": "#/components/schemas/WorldSpec"
          },
          "rig": {
            "$ref": "#/components/schemas/RigSpec"
          },
          "environment": {
            "$ref": "#/components/schemas/EnvironmentSpec"
          },
          "plan": {
            "$ref": "#/components/schemas/CapturePlan"
          }
        },
        "required": [
          "schema_version",
          "kind",
          "world",
          "rig",
          "environment",
          "plan"
        ]
      },
      "CaptureJobError": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "enum": [
              "worker_crash",
              "worker_timeout",
              "worker_context_lost",
              "worker_interrupted",
              "worker_error"
            ]
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1024
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "CaptureResult": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "protocol": {
            "const": "capture-worker.v1"
          },
          "capture_id": {
            "$ref": "#/components/schemas/Id"
          },
          "sequence_id": {
            "$ref": "#/components/schemas/Id"
          },
          "tick_s": {
            "type": "number",
            "minimum": 0,
            "maximum": 1000000000
          },
          "width": {
            "type": "integer",
            "minimum": 1,
            "maximum": 8192
          },
          "height": {
            "type": "integer",
            "minimum": 1,
            "maximum": 8192
          },
          "renderer": {
            "type": "string",
            "minLength": 1,
            "maxLength": 512
          },
          "device": {
            "type": "string",
            "minLength": 1,
            "maxLength": 512
          },
          "browser_channel": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "elapsed_ms": {
            "type": "number",
            "minimum": 0
          },
          "render_elapsed_ms": {
            "type": "number",
            "minimum": 0
          },
          "node_rss_bytes": {
            "type": "integer",
            "minimum": 0
          },
          "browser_heap_bytes": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "gpu_memory_bytes": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "resident_chunks": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64
            },
            "minItems": 1,
            "maxItems": 256,
            "uniqueItems": true
          },
          "instance_ids": {
            "type": "object",
            "propertyNames": {
              "$ref": "#/components/schemas/Id"
            },
            "additionalProperties": {
              "type": "integer",
              "minimum": 1,
              "maximum": 16777215
            },
            "maxProperties": 100000
          },
          "ir_calibration": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IrCalibration"
              },
              {
                "type": "null"
              }
            ]
          },
          "lidar_calibration": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LidarCalibration"
              },
              {
                "type": "null"
              }
            ]
          },
          "lidar_point_count": {
            "type": "integer",
            "minimum": 0,
            "maximum": 8388608
          },
          "artifacts": {
            "$ref": "#/components/schemas/CaptureArtifacts"
          }
        },
        "required": [
          "protocol",
          "capture_id",
          "sequence_id",
          "tick_s",
          "width",
          "height",
          "renderer",
          "device",
          "browser_channel",
          "elapsed_ms",
          "render_elapsed_ms",
          "node_rss_bytes",
          "browser_heap_bytes",
          "gpu_memory_bytes",
          "resident_chunks",
          "instance_ids",
          "ir_calibration",
          "lidar_calibration",
          "lidar_point_count",
          "artifacts"
        ]
      },
      "CaptureJob": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "schema_version": {
            "const": "lab.v1"
          },
          "kind": {
            "const": "CaptureJob"
          },
          "job_id": {
            "$ref": "#/components/schemas/Id"
          },
          "capture_id": {
            "$ref": "#/components/schemas/Id"
          },
          "state": {
            "enum": [
              "queued",
              "running",
              "cancelling",
              "cancelled",
              "succeeded",
              "failed"
            ]
          },
          "progress": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CaptureResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CaptureJobError"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "schema_version",
          "kind",
          "job_id",
          "capture_id",
          "state",
          "progress",
          "result",
          "error",
          "created_at",
          "updated_at"
        ]
      },
...
      "Capability": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "available": {
            "type": "boolean"
          },
          "reason": {
            "anyOf": [
              {
                "enum": [
                  "not_implemented",
                  "no_checkpoint",
                  "device_unavailable"
                ]
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "available",
          "reason"
        ],
        "allOf": [
          {
            "if": {
              "properties": {
                "available": {
                  "const": true
                }
              }
            },
            "then": {
              "properties": {
                "reason": {
                  "type": "null"
                }
              }
            },
            "else": {
              "properties": {
                "reason": {
                  "enum": [
                    "not_implemented",
                    "no_checkpoint",
                    "device_unavailable"
                  ]
                }
              }
            }
          }
        ]
      },
      "Capabilities": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "schema_version": {
            "const": "lab.v1"
          },
          "kind": {
            "const": "Capabilities"
          },
          "service_version": {
            "const": "0.1.0"
          },
          "contracts_version": {
            "const": "lab.v1"
          },
          "capture_renderer": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1,
                "maxLength": 256
              },
              {
                "type": "null"
              }
            ]
          },
          "device": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1,
                "maxLength": 256
              },
              {
                "type": "null"
              }
            ]
          },
          "sensors": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "rgb": {
                "$ref": "#/components/schemas/Capability"
              },
              "ir": {
                "$ref": "#/components/schemas/Capability"
              },
              "lidar": {
                "$ref": "#/components/schemas/Capability"
              }
            },
            "required": [
              "rgb",
              "ir",
              "lidar"
            ]
          },
          "services": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "worlds": {
                "$ref": "#/components/schemas/Capability"
              },
              "capture": {
                "$ref": "#/components/schemas/Capability"
              },
              "datasets": {
                "$ref": "#/components/schemas/Capability"
              },
              "training": {
                "$ref": "#/components/schemas/Capability"
              },
              "inference": {
                "$ref": "#/components/schemas/Capability"
              }
            },
            "required": [
              "worlds",
              "capture",
              "datasets",
              "training",
              "inference"
            ]
          },
          "checkpoints": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Id"
            },
            "minItems": 0,
            "maxItems": 1000
          }
        },
        "required": [
          "schema_version",
          "kind",
          "service_version",
          "contracts_version",
          "capture_renderer",
          "device",
          "sensors",
          "services",
          "checkpoints"
        ]
      },
      "ApiError": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "schema_version": {
            "const": "lab.v1"
          },
          "kind": {
            "const": "ApiError"
          },
          "code": {
            "enum": [
              "not_found",
              "method_not_allowed",
              "validation_error",
              "internal_error"
            ]
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1024
          },
          "request_id": {
            "$ref": "#/components/schemas/Id"
          },
          "job_id": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Id"
              },
              {
                "type": "null"
              }
            ]
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": false,
              "properties": {
                "path": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 512
                },
                "message": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 1024
                }
              },
              "required": [
                "path",
                "message"
              ]
            },
            "minItems": 0,
            "maxItems": 100
          }
        },
        "required": [
          "schema_version",
          "kind",
          "code",
          "message",
          "request_id",
          "job_id",
          "details"
        ]
      }
    },
...
  }
}